Move quadratic_equivalences
