+++ constraint-solver/src/lib.rs
@@ -3,7 +3,6 @@
 pub mod constraint_system;
 pub mod effect;
 pub mod indexed_constraint_system;
-mod quadratic_equivalences;
 pub mod quadratic_symbolic_expression;
 pub mod range_constraint;
 pub mod solver;
