+++ constraint-solver/src/solver/quadratic_equivalences.rs
